Adding Achievements, have the tracking set up need to write up a screen to display them.

+++ app/src/main/java/online/cagocapps/walkingthedog/MainActivity.java
@@ -309,7 +309,7 @@ public class MainActivity extends AppCompatActivity implements SharedPreferences
             else tv.setText("0");
             tv = (TextView) findViewById(R.id.ave_mph_value);
             double math;
-            double time = (double) totTimes;
+            double time = totTimes;
             if (totTimes != 0) {
                 math = (totDist / (time / 60));
                 if (math < .01) math = 0;
